+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ## Fixed
 - allow scanning of certain qr code types on welcome screen (account, url and text)
 - fix selecting chat after forwarding to it
+- fixed bug where unread badge on app icon was not updated immediately
 
 ## [1.32.1] - 2022-08-18